+++ rs/tests/consensus/BUILD.bazel
@@ -47,6 +47,7 @@ rust_library(
     crate_name = "ic_consensus_system_test_catch_up_test_common",
     deps = [
         # Keep sorted.
+        "//rs/tests/consensus/utils",
         "//rs/tests/driver:ic-system-test-driver",
         "//rs/types/types",
         "@crate_index//:anyhow",
