+++ rs/tests/consensus/catch_up_test_common/Cargo.toml
@@ -9,6 +9,7 @@ documentation.workspace = true
 [dependencies]
 anyhow = { workspace = true }
 futures = { workspace = true }
+ic_consensus_system_test_utils = { path = "../utils" }
 ic-system-test-driver = { path = "../../driver" }
 ic-types = { path = "../../../types/types" }
 slog = { workspace = true }
