+++ rs/tests/consensus/catch_up_test_common/src/lib.rs
@@ -13,17 +13,18 @@ use ic_system_test_driver::{
     driver::{
         test_env::TestEnv,
         test_env_api::{
-            HasPublicApiUrl, HasTopologySnapshot, HasVm, IcNodeContainer, IcNodeSnapshot,
+            HasPublicApiUrl, HasTopologySnapshot, HasVm, IcNodeContainer,
             READY_WAIT_TIMEOUT, RETRY_BACKOFF,
         },
     },
     util::{block_on, MetricsFetcher},
 };
 use ic_types::Height;
+use ic_consensus_system_test_utils::node::{await_node_certified_height, get_node_certified_height}; 
 
-use anyhow::{anyhow, bail};
+use anyhow::bail;
 use futures::join;
-use slog::{info, Logger};
+use slog::info;
 use std::time::Duration;
 
 const PROMETHEUS_SCRAPE_INTERVAL: Duration = Duration::from_secs(5);
@@ -66,7 +67,7 @@ fn test(env: TestEnv, expect_catch_up: bool) {
         malicious_node.malicious_behavior().unwrap()
     );
 
-    let slow_node_certified_height = get_certified_height(&malicious_node, log.clone()).get();
+    let slow_node_certified_height = get_node_certified_height(&malicious_node, log.clone()).get();
 
     let slow_node_shut_down_height = DKG_INTERVAL * (1 + slow_node_certified_height / DKG_INTERVAL);
     info!(log, "Wait one DKG interval, then shut down the slow node");
@@ -221,44 +222,3 @@ fn test(env: TestEnv, expect_catch_up: bool) {
     });
 }
 
-pub fn await_node_certified_height(node: &IcNodeSnapshot, target_height: Height, log: Logger) {
-    ic_system_test_driver::retry_with_msg!(
-        format!(
-            "check if node {} is at height {}",
-            node.node_id, target_height
-        ),
-        log,
-        READY_WAIT_TIMEOUT,
-        RETRY_BACKOFF,
-        || {
-            node.status()
-                .and_then(|response| match response.certified_height {
-                    Some(height) if height > target_height => Ok(()),
-                    Some(height) => bail!(
-                        "Target height not yet reached, height: {}, target: {}",
-                        height,
-                        target_height
-                    ),
-                    None => bail!("Certified height not available"),
-                })
-        }
-    )
-    .expect("The node did not reach the specified height in time")
-}
-
-pub fn get_certified_height(node: &IcNodeSnapshot, log: Logger) -> Height {
-    ic_system_test_driver::retry_with_msg!(
-        format!("get certified height of node {}", node.node_id),
-        log,
-        READY_WAIT_TIMEOUT,
-        RETRY_BACKOFF,
-        || {
-            node.status().and_then(|response| {
-                response
-                    .certified_height
-                    .ok_or_else(|| anyhow!("Certified height not available"))
-            })
-        }
-    )
-    .expect("Should be able to retrieve the certified height")
-}
